refactor: useMyContractsList 생성 후 contract 페이지에서 사용
